refactor: change isOpen option name to opened
