+++ examples/example01-code.html
@@ -18,7 +18,7 @@
           { value: 'apple', text: 'Apple', selected: true },
           { value: 'banana', text: 'Banana' },
         ],
-        isOpen: true,
+        autofocus: true,
         // value: 'banana'
       });
     </script>
