+++ src/store/create.ts
@@ -1,8 +1,12 @@
-import { Store } from '@t/store';
+import { Store, RenderState } from '@t/store';
 import { observable } from '../helper/observable';
 import { SelectOptions } from '@t/options';
-import { create as createRenderState } from './renderState';
 import { create as createOptions } from './data';
+
+function createRenderState(options: SelectOptions) {
+  const opened = options.autofocus ?? false;
+  return observable({ hoveredKey: null, opened } as RenderState);
+}
 
 export function createStore(id: number, _options: SelectOptions): Store {
   const renderState = createRenderState(_options);
