+++ types/store/index.d.ts
@@ -1,5 +1,4 @@
 import { Data } from './data';
-import { RenderState } from './renderState';
 
 export type ComponentId = number;
 
@@ -7,4 +6,9 @@ export interface Store {
   readonly id: ComponentId;
   renderState: RenderState;
   data: Data;
+}
+
+export interface RenderState {
+  opened: boolean;
+  hoveredKey: string | null;
 }
