--- types/store/renderState.d.ts
@@ -1,4 +1,0 @@
-export interface RenderState {
-  isOpen: boolean;
-  hoveredKey: string | null;
-}
